BUGFIX: entrypoint.sh and Hugo Version
The original code made the assumption that keeping the rightmost 6 characters would always work. This fails when Hugo's maintainers rolled into v0.100.0, adding another digit. Since the intent is to siphon off the "v", this fix does just that, so the version number digits can be any length and it will not break.